refactor: change hash structure
