+++ src/modules/handlers/common_handler.js
@@ -1,7 +1,7 @@
 import { exitApp } from "./../operation/exit.js";
 import { parseCommand } from "./../utils/command_utils.js";
 import { system_info } from "./os_handler.js";
-import { hash_info } from "./hash_handler.js";
+import { handleHashCommand } from "./hash_handler.mjs";
 import { handleNwdCommand } from "./nwd_handler.js";
 import { handleFileCommand } from "./file_ops_handler.js";
 import { handleZipCommand } from "./zip_ops_handler.js"
@@ -20,7 +20,7 @@ const handleCommand = async (command, username) => {
       break;
 
     case 'hash':
-      await hash_info(parsedCommand.args);
+      await handleHashCommand(parsedCommand.args);
       break;
 
     case 'ls':
