+++ src/modules/handlers/hash_handler.mjs
@@ -1,13 +1,11 @@
-import { calculateHash } from "./../operation/hash.js";
+import { calculateHash } from "../hash/hash.mjs";
 
-const hash_info = async (args) => {
+export const handleHashCommand = async (args) => {
     if (args.length === 0) {
         console.log("the 'hash' command must contains at least one argument.");
         return;
     }
     let filePath = args[0];
 
-   await calculateHash(filePath);
+    await calculateHash(filePath);
 };
-
-export { hash_info }
